+++ src/coordinax/_src/vectors/base/mixins.py
@@ -33,7 +33,7 @@ class AvalMixin:
 
         >>> vec = cx.vecs.CartesianPos1D.from_([1], "m")
         >>> vec.aval()
-        ConcreteArray([1.], dtype=float32)
+        ConcreteArray([1], dtype=int32)
 
         >>> vec = cx.vecs.RadialPos.from_([1], "m")
         >>> vec.aval()
